+++ src/utils/output_json.h
@@ -35,6 +35,9 @@ rapidjson::Value to_json(const Step& s,
                          bool report_distances,
                          rapidjson::Document::AllocatorType& allocator);
 
+rapidjson::Value to_json(const CostBreakdown& breakdown,
+                         rapidjson::Document::AllocatorType& allocator);
+
 rapidjson::Value to_json(const Location& loc,
                          rapidjson::Document::AllocatorType& allocator);